+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# Version 1.8.0
+
+Compatible with Kimai version 2.38.0 or higher.
+
+- Added break field and validator
+- Updated validator
+  - Duration cannot be longer than 24 hours
+  - Fixed budget validator
+
+
 # Version 1.7.0
 
 Compatible with Kimai version 2.36.0 or higher.